[#3327] Upgrade to Quarkus 2.10.2
Fixes #3327
