+++ bom/pom.xml
@@ -57,18 +57,13 @@
     <logback.version>1.2.9</logback.version>
     <mongodb-image.name>mongo:4.4</mongodb-image.name>
     <native.image.name>quay.io/quarkus/quarkus-micro-image:1.0</native.image.name>
-    <!--
-      The value of this property should always be the Java 17 based variant of the image name defined
-      in the Quarkus platform definition's POM file of the Quarkus version being used, e.g.
-      https://github.com/quarkusio/quarkus-platform/blob/2.8.2.Final/generated-platform-project/quarkus/properties/pom.xml
-    -->
-    <native.builder-image.name>quay.io/quarkus/ubi-quarkus-mandrel:22.0-java17</native.builder-image.name>
+    <native.builder-image.name>mandrel</native.builder-image.name>
     <netty.version>4.1.74.Final</netty.version>
-    <opentelemetry-collector.image.name>otel/opentelemetry-collector:0.49.0</opentelemetry-collector.image.name>
+    <opentelemetry-collector.image.name>otel/opentelemetry-collector:0.55.0</opentelemetry-collector.image.name>
     <postgresql-image.name>postgres:12-alpine</postgresql-image.name>
     <proton.version>0.33.10</proton.version>
     <qpid-jms.version>1.4.1</qpid-jms.version>
-    <quarkus.platform.version>2.8.3.Final</quarkus.platform.version>
+    <quarkus.platform.version>2.10.2.Final</quarkus.platform.version>
     <slf4j.version>1.7.36</slf4j.version>
     <spring-security-crypto.version>5.7.0</spring-security-crypto.version>
     <truth.version>1.1.3</truth.version>
