+++ service-base/src/main/resources/META-INF/native-image/org.eclipse.hono/hono-service-base/native-image.properties
@@ -12,5 +12,4 @@
 # use Base64 encoder/decoder that is compatible with vert.x 3
 Args = -H:ResourceConfigurationResources=${.}/resources-config.json \
   --initialize-at-run-time=io.netty.internal.tcnative.SSL \
-  --allow-incomplete-classpath \
   -Dvertx.json.base64=legacy
