+++ services/command-router/src/main/java/org/eclipse/hono/commandrouter/AdapterInstanceStatusService.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 Contributors to the Eclipse Foundation
+ * Copyright (c) 2021, 2022 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -13,46 +13,11 @@
 
 package org.eclipse.hono.commandrouter;
 
-import java.util.Collection;
-import java.util.Set;
-
 import org.eclipse.hono.deviceconnection.infinispan.client.AdapterInstanceStatusProvider;
-import org.eclipse.hono.util.AdapterInstanceStatus;
 import org.eclipse.hono.util.Lifecycle;
-
-import io.vertx.core.Future;
 
 /**
  * A service for determining the status of adapter instances.
  */
 public interface AdapterInstanceStatusService extends AdapterInstanceStatusProvider, Lifecycle {
-
-    UnknownStatusProvidingService UNKNOWN_STATUS_PROVIDING_SERVICE = new UnknownStatusProvidingService();
-
-    /**
-     * Status service that always returns the {@link AdapterInstanceStatus#UNKNOWN} status.
-     */
-    class UnknownStatusProvidingService implements AdapterInstanceStatusService {
-
-        @Override
-        public AdapterInstanceStatus getStatus(final String adapterInstanceId) {
-            return AdapterInstanceStatus.UNKNOWN;
-        }
-
-        @Override
-        public Future<Set<String>> getDeadAdapterInstances(
-                final Collection<String> adapterInstanceIds) {
-            return Future.succeededFuture(Set.of());
-        }
-
-        @Override
-        public Future<Void> start() {
-            return Future.succeededFuture();
-        }
-
-        @Override
-        public Future<Void> stop() {
-            return Future.succeededFuture();
-        }
-    }
 }
